+++ efc/inc/EFileOutputStream.hh
@@ -186,26 +186,6 @@ public:
     virtual void write(const void *b, int len) THROWS(EIOException);
     virtual void write(const char *s) THROWS(EIOException);
     virtual void write(int b) THROWS(EIOException);
-	
-	/**
-     * Flushes this output stream and forces any buffered output bytes 
-     * to be written out. The general contract of <code>flush</code> is 
-     * that calling it is an indication that, if any bytes previously 
-     * written have been buffered by the implementation of the output 
-     * stream, such bytes should immediately be written to their 
-     * intended destination.
-     * <p>
-     * If the intended destination of this stream is an abstraction provided by
-     * the underlying operating system, for example a file, then flushing the
-     * stream guarantees only that bytes previously written to the stream are
-     * passed to the operating system for writing; it does not guarantee that
-     * they are actually written to a physical device such as a disk drive.
-     * <p>
-     * The <code>flush</code> method of <code>OutputStream</code> does nothing.
-     *
-     * @exception  IOException  if an I/O error occurs.
-     */
-    virtual void flush() THROWS(EIOException);
     
 	/**
      * Closes this file output stream and releases any system resources