+++ efc/inc/EIterable.hh
@@ -21,6 +21,30 @@ namespace efc {
  * @since 1.5
  */
 
+#ifdef CPP11_SUPPORT
+template<typename T>
+class RangeIter {
+public:
+	RangeIter() {}
+	RangeIter(sp<EIterator<T> > i) : i(i) {}
+
+	bool operator!=(const RangeIter<T>& other) const {
+		return i->hasNext();
+	}
+
+	const T operator*() {
+		return i->next();
+	}
+
+	const RangeIter& operator++() {
+		return *this;
+	}
+
+private:
+	sp<EIterator<T> > i;
+};
+#endif
+
 template<typename T>
 interface EIterable : virtual public EObject
 {
@@ -32,6 +56,15 @@ interface EIterable : virtual public EObject
      * @return an Iterator.
      */
     virtual sp<EIterator<T> > iterator(int index=0) = 0;
+
+#ifdef CPP11_SUPPORT
+    RangeIter<T> begin() {
+    	return RangeIter<T>(iterator());
+    }
+    RangeIter<T> end() {
+    	return RangeIter<T>();
+    }
+#endif
 };
 
 } /* namespace efc */