+++ test/testc11.cpp
@@ -5,6 +5,106 @@
 
 #ifdef CPP11_SUPPORT
 
+#ifndef WIN32
+#include <mutex>
+#endif
+
+static void thread_worker()
+{
+    LOG("thread_worker running...");
+}
+
+static void thread_worker2(int v)
+{
+    LOG("thread_worker2 [v=%d] running...", v);
+}
+
+#define LOCKED_WORK_THREADS 10
+#define MAX_LOCKED_COUNTE 10000000
+static long locked_counter = 0L;
+static EReentrantLock locked_counter_lock;
+#ifndef WIN32
+static std::mutex locked_counter_stdmutex;
+static pthread_mutex_t locked_counter_mutex = PTHREAD_MUTEX_INITIALIZER;
+#endif
+static void test_lock_benchmark() {
+	class Thread: public EThread {
+	public:
+		virtual void run() {
+			while (true) {
+				//1. use ELockPool
+				//mac: 10 thread run 10000000 times, cost 1494 ms
+				//linux: 10 thread run 10000000 times, cost 1737 ms
+				//linux2: 10 thread run 10000000 times, cost 3554 ms/1 thread run 10000000 times, cost 160 ms/2 thread run 10000000 times, cost 2582 ms
+				if (0) {
+					SCOPED_SLOCK5(&locked_counter) {
+						locked_counter++;
+						if (locked_counter > MAX_LOCKED_COUNTE) {
+							break;
+						}
+					}}
+				}
+
+				//2. use EReentrantLock
+				//mac: 10 thread run 10000000 times, cost 3276 ms
+				//linux: 10 thread run 10000000 times, cost 1512 ms
+				//linux2: 10 thread run 10000000 times, cost 1588 ms/1 thread run 10000000 times, cost 595 ms/2 thread run 10000000 times, cost 6031 ms
+				if (1) {
+					SYNCBLOCK(&locked_counter_lock) {
+						locked_counter++;
+						if (locked_counter > MAX_LOCKED_COUNTE) {
+							break;
+						}
+					}}
+				}
+
+#ifndef WIN32
+				//3. use std::mutex
+				if (0) {
+					std::lock_guard<std::mutex> lock(locked_counter_stdmutex);
+					locked_counter++;
+					if (locked_counter > MAX_LOCKED_COUNTE) {
+						break;
+					}
+				}
+
+				//4. pthread mutex
+				//mac: 10 thread run 10000000 times, cost 46415 ms
+				//linux: 10 thread run 10000000 times, cost 975 ms
+				//linux2: 10 thread run 10000000 times, cost 1096 ms/1 thread run 10000000 times, cost 241 ms/2 thread run 10000000 times, cost 1047 ms
+				if (0) {
+					pthread_mutex_lock( &locked_counter_mutex );
+					locked_counter++;
+					if (locked_counter > MAX_LOCKED_COUNTE) {
+						pthread_mutex_unlock( &locked_counter_mutex );
+						break;
+					}
+					pthread_mutex_unlock( &locked_counter_mutex );
+				}
+#endif
+			}
+		}
+	};
+
+	llong t1 = ESystem::currentTimeMillis();
+
+	EArrayList<sp<Thread> > arrs;
+	for (int i=0; i<LOCKED_WORK_THREADS; i++) {
+		sp<Thread> t = new Thread();
+		arrs.add(t);
+		t->start();
+	}
+
+	for (int i=0; i<LOCKED_WORK_THREADS; i++) {
+		arrs.getAt(i)->join();
+	}
+
+	llong t2 = ESystem::currentTimeMillis();
+
+	LOG("%d thread run %ld times, cost %ld ms\nper second op times: %f",
+			LOCKED_WORK_THREADS, MAX_LOCKED_COUNTE, t2 - t1, ((double)MAX_LOCKED_COUNTE)/(t2-t1)*1000);
+}
+
 static void test_scopeExit() {
 	EString* s1 = new EString("xxx");
 	ON_SCOPE_EXIT(delete s1;);
@@ -122,7 +222,7 @@ static void test_finally() {
 static void test_threadx() {
 	ESynchronizeable sync;
 
-	sp<EThread> ths1 = EThread::executeX([&]() {
+	auto ths1 = EThread::executeX([&]() {
 		//EThread::sleep(1000); //if wait after notify then will blocked forever.
 		SYNCHRONIZED(&sync) {
 			sync.wait();
@@ -130,7 +230,7 @@ static void test_threadx() {
 	});
 //	ths1->start(); //exception: Already started.
 
-	sp<EThread> ths2 = EThread::executeX([&]() {
+	auto ths2 = EThread::executeX([&]() {
 		EThread::sleep(1000);
 		SYNCHRONIZED(&sync) {
 			LOG("at here.");
@@ -142,13 +242,19 @@ static void test_threadx() {
 	ths1->join();
 	ths2->join();
 
+	auto ths3 = EThread::executeX(thread_worker);
+	ths3->join();
+
+	auto ths4 = EThread::executeX(std::bind(thread_worker2, 4));
+	ths4->join();
+
 	LOG("end of test_threadx().");
 }
 
 static void test_timerx() {
 	ETimer timer;
 
-	sp<ETimerTask> task = timer.scheduleX((llong)0, 100, []() {
+	auto task = timer.scheduleX((llong)0, 100, []() {
 		LOG("timer runging...");
 	});
 
@@ -247,6 +353,26 @@ static void test_socketpair() {
 #endif
 }
 
+static void test_foreach() {
+#if !defined(_MSC_VER) || (_MSC_VER>=1800) //VS2013
+	EArray<EString*> arr;
+	arr.add(new EString("1"));
+	arr.add(new EString("2"));
+	arr.add(new EString("3"));
+	for (auto s : arr) {
+		LOG("s=%s", s->c_str());
+	}
+
+	EConcurrentHashMap<int, EInteger> hm;
+	hm.put(1, new EInteger(1));
+	hm.put(2, new EInteger(2));
+	hm.put(3, new EInteger(3));
+	for (auto i : *hm.entrySet()) {
+		LOG("i=%d", i->getValue()->intValue());
+	}
+#endif
+}
+
 MAIN_IMPL(testc11) {
 	printf("main()\n");
 
@@ -256,18 +382,20 @@ MAIN_IMPL(testc11) {
 
 	do {
 		try {
+//			test_lock_benchmark();
 //			test_scopeExit();
 //			test_finally();
-//			test_threadx();
-			test_timerx();
+			test_threadx();
+//			test_timerx();
 //			test_executors();
 //			test_socketpair();
+//			test_foreach();
 		} catch (EException& e) {
 			LOG("exception: %s", e.getMessage());
 		} catch (...) {
 			LOG("Catched a exception.");
 		}
-	} while (1);
+	} while (0);
 
 	LOG("exit...");